test(symbolic): split parity corpora
