+++ crates/forge/tests/cli/test_cmd/mod.rs
@@ -19,8 +19,21 @@ mod repros;
 mod spec;
 mod symbolic;
 mod symbolic_conformance;
-mod symbolic_fuzz_parity;
+mod symbolic_engine_capabilities;
+mod symbolic_helpers;
 mod symbolic_limits;
+mod symbolic_parity_crytic;
+mod symbolic_parity_devdacian;
+mod symbolic_parity_echidna;
+mod symbolic_parity_halmos;
+mod symbolic_parity_hevm;
+mod symbolic_parity_ityfuzz;
+mod symbolic_parity_kevm;
+mod symbolic_parity_manticore;
+mod symbolic_parity_medusa;
+mod symbolic_parity_openzeppelin;
+mod symbolic_parity_scribble;
+mod symbolic_parity_swc;
 mod table;
 mod trace;
